Fix chemDoodle redeclared issue
This problem arises because I injected the js script whenever the component is being mounted/unmounted, so
everytime we navigate to a new page or load a new question, the script will be loaded again, causing the chemdoodle to be
duplicated. To fix this, we can load the script in the main layout page, so the JS logic will only be loaded once in the
entire app life cycle.